requirements.txt
appnope==0.1.3
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.16.3
fire==0.5.0
fonttools==4.39.0
icon-font-to-png==0.4.1
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
ipykernel==6.21.3
ipython==8.11.0
jedi==0.18.2
jsonschema==4.17.3
jupyter_client==8.0.3
jupyter_core==5.2.0
kiwisolver==1.4.4
matplotlib==3.7.1
matplotlib-inline==0.1.6
nbformat==5.7.3
nest-asyncio==1.5.6
numpy==1.24.2
openpyxl==3.1.2
packaging==23.0
palettable==3.3.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.1
plotly==5.13.1
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pythumb==1.1.0
pytz==2022.7.1
pyzmq==25.0.0
requests==2.28.2
six==1.16.0
stack-data==0.6.2
stylecloud==0.5.2
tenacity==8.2.2
termcolor==2.2.0
tinycss==0.4
tornado==6.2
traitlets==5.9.0
urllib3==1.26.15
wcwidth==0.2.6
wordcloud==1.8.2.2
zipp==3.15.0
